+++ sdk/android/src/main/java/com/reactnativehypertracksdk/common/SdkMethod.kt
@@ -16,10 +16,12 @@ internal enum class SdkMethod {
     getLocation,
     getMetadata,
     getName,
+    getWorkerHandle,
     locate,
     setDynamicPublishableKey,
     setIsAvailable,
     setIsTracking,
     setMetadata,
     setName,
+    setWorkerHandle,
 }