+++ sdk/src/HyperTrack/data_types/internal/WorkerHandle.ts
@@ -1,0 +1,4 @@
+export type WorkerHandle = {
+  type: 'workerHandle';
+  value: string;
+};