+++ CHANGELOG.md
@@ -5,6 +5,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [13.4.0] - 2024-06-13
+
+### Added
+
+- New `setWorkerHandle` and `getWorkerHandle` can be used to identify workers
+  - We observed our customers identify worker devices via `HyperTrack.metadata`, so we decided to make it a first class citizen in our API.
+  - If you previously used `metadata` to identify workers, we suggest using `workerHandle` for this purpose instead.
+
+### Changed
+
+- Updated HyperTrack SDK Android to [7.6.0](https://github.com/hypertrack/sdk-android/releases/tag/7.6.0)
+
 ## [13.3.5] - 2024-06-05
 
 ### Changed
@@ -856,3 +868,4 @@ Initial release.
 [13.3.3]: https://github.com/hypertrack/sdk-react-native/releases/tag/13.3.3
 [13.3.4]: https://github.com/hypertrack/sdk-react-native/releases/tag/13.3.4
 [13.3.5]: https://github.com/hypertrack/sdk-react-native/releases/tag/13.3.5
+[13.4.0]: https://github.com/hypertrack/sdk-react-native/releases/tag/13.4.0